Problem Set A9: Bug fixed

+++ ProblemsetA/a9_Submission_is_All_You_Need_II.c
@@ -4,9 +4,9 @@ int main() {
     int t;
     scanf("%d", &t);
     while (t--) {
-        int n;
-        scanf("%d", &n);
-        printf("%d\n", n);
+        long long int n;
+        scanf("%lld", &n);
+        printf("%lld\n",n);
     }
     return 0;
 }
